Further factor out common logic.
